+++ plugins/php/versions/82/install.sh
@@ -31,7 +31,7 @@ if [ ! -d $sourcePath/php/php${PHP_VER} ];then
 
 	# ----------------------------------------------------------------------- #
 	# 中国优化安装
-	cn=$(curl -fsSL -m 10 -s https://ipinfo.io/json | grep "\"country\": \"CN\"")
+	cn=$(curl -fsSL -k -m 10 -s https://ipinfo.io/json | grep "\"country\": \"CN\"")
 	LOCAL_ADDR=common
 	if [ ! -z "$cn" ] || [ "$?" == "0" ] ;then
 		LOCAL_ADDR=cn
@@ -54,9 +54,14 @@ if [ ! -d $sourcePath/php/php${PHP_VER} ];then
 		if [ "${md5_file}" != "${md5_file_ok}" ]; then
 			echo "PHP${version} 下载文件不完整,重新安装"
 			rm -rf $sourcePath/php/php-${version}.tar.xz
+			exit 1
 		fi
 	fi
 	
+	if [ ! -f $sourcePath/php/php-${version}.tar.xz ]; then
+		echo "PHP source file missing!"
+		exit 1
+	fi
 	cd $sourcePath/php && tar -Jxf $sourcePath/php/php-${version}.tar.xz
 	mv $sourcePath/php/php-${version} $sourcePath/php/php${PHP_VER}
 fi
